% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom_enhanced_kwds.R
\name{create_custom_enhanced_keywords}
\alias{create_custom_enhanced_keywords}
\title{Create additional keywords by creating your own combinations of queries and
    choosen vectors.}
\usage{
create_custom_enhanced_keywords(queries, suffix_vec = NULL)
}
\arguments{
\item{queries}{vector of character strings. 'Bare' queries that will feed
the Google Suggest.}

\item{suffix_vec}{vector of character strings. Optional. It will generate
additional combinations of keywords and prefix_vec contents will be
placed AFTER the queries. Defaults to NULL}

\item{prefix_vec}{vector of character strings. Optional. It will generate
additional combinations of keywords and prefix_vec contents will be
placed BEFORE the queries. Defaults to NULL.}
}
\value{
vector of character strings with all combinations of queries and
    suffixes/prefixes
}
\description{
Create additional keywords by creating your own combinations of queries and
    choosen vectors.
}
\examples{
\dontrun{
  create_custom_enhanced_keywords(
    queries = "mtcars",
    suffix_vec = letters,
    prefix_vec = c(0:9)
  )
}
}
