% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_basic_enhanced_kwds.R
\name{create_enhanced_keywords}
\alias{create_enhanced_keywords}
\title{Create additional keywords by creating combinations of queries and letters
   and digits.}
\usage{
create_enhanced_keywords(queries)
}
\arguments{
\item{queries}{vector of character strings. 'Bare' queries that will feed
the Google Suggest.}
}
\value{
vector of character strings with all combinations of queries and
    alphabet letters and digits
}
\description{
Create additional keywords by creating combinations of queries and letters
   and digits.
}
\examples{
\dontrun{
  create_enhanced_keywords(
    queries = c("mtcars", "iris")
  )
}
}
