% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_kwds.R
\name{suggest_keywords}
\alias{suggest_keywords}
\title{Download suggested keywords for your keywords vector}
\usage{
suggest_keywords(queries, lang = "en", interval = 0, enhanced = T)
}
\arguments{
\item{queries}{vector of character strings. The queries that will be used to
create suggestions}

\item{lang}{character string. The language for which the suggestions will be
created, formatted like: "en", "es", "fr", "de", "pl" etc.}

\item{interval}{non-negative number. How many seconds must the function wait
after the call to API? Defaults to 0, adjust if exceeded API limits
encountered}

\item{enhanced}{logical. Do you want to extract more suggestions by adding
letters and digits after input queries? Defaults to TRUE}
}
\value{
vector of keywords suggestion
}
\description{
Download suggested keywords for your keywords vector
}
\examples{
\dontrun{
  kwds <- suggest_keywords(
    queries = c("mtcars", "iris"),
    lang = "en",
    interval = 1,
    enhanced = T
  )
}
}
