% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_single_kwd.R
\name{suggest_single_keyword}
\alias{suggest_single_keyword}
\title{Download suggested keywords for a single input query}
\usage{
suggest_single_keyword(query, lang = "en", interval = 0)
}
\arguments{
\item{query}{character string. The query that will be used to create suggestions}

\item{lang}{character string. The language for which the suggestions will be
created, formatted like: "en", "es", "fr", "de", "pl" etc.}

\item{interval}{non-negative number. How many seconds must the function wait
after the call to API? Defaults to 0, adjust if exceeded API limits
encountered}
}
\value{
single column data frame with suggested queries
}
\description{
Download suggested keywords for a single input query
}
\examples{
\dontrun{
  kwds <- suggest_single_keyword(
    queries = "mtcars",
    lang = "en",
    interval = 1
  )
}
}
